+++ src/main/kotlin/org/radarbase/push/integration/fitbit/service/subscription/SubscriptionRequest.kt
@@ -1,4 +1,4 @@
-package org.radarbase.push.integration.fitbit.subscription
+package org.radarbase.push.integration.fitbit.service.subscription
 
 import okhttp3.Request
 import org.radarbase.push.integration.common.user.User
